Diseño de interfaz, implementación de método listar

+++ dominio/Articulo.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Data.SqlTypes;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -16,8 +17,6 @@ namespace dominio
         public Marca Marca { get; set; }
         public Categoria Categoria { get; set; }
         public string ImagenUrl { get; set; }
-        public float Precio { get; set; }
-
-        // constructor y metodos
+        public Decimal Precio { get; set; }
     }
 }
